+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# v0.2.0
+
+* doc: update azure readme
+* upd: clarify time delta calculation
+* add: more aggressive ticker to start collections
+* add: log collection duration
+* upd: remove dead code
+* fix: dynamic dimensions from call param
+* fix: catch zero instances for a given region
+* add: cloud vendor specific check submodule
+* upd: add region debug message
+* upd: reduce ticker drift to 2sec
+* upd: promote duration message to Info
+* upd: only print submission in debug mode
+* upd: print response body as raw json
+* fix: verify returned check bundle has a submission url
+* upd: dependencies
+* add: use service specific check submodules
+* upd: add interval logging on client start
+* upd: add additional aws colleciton logging
+* upd: adjust aws collection interval delta calculation
+* upd: stricter linting
+* doc: update install instructions (correct repository for releases)
+
 # v0.1.1
 
 * fix: release to circonus-labs repository